Fix bug when Twitter returns an HTML error page rather than XML with an error node in it
Apparently Twitter returns errors in two ways: 1) by returning an HTML
error page with a <title> containing the error, and 2) by returning an
XML result with an <error> node in it containing the error. My code was
checking for (2) but not for (1). So what happened was the code thought
the returned XML was valid since it didn't contain an <error> node.

Unfortunately since the returned XML was just a Twitter HTML page and
didn't contain any statuses, parsing it into statuses resulted in a
runtime error. The fix was to check for a result which contains a
<title> node and if it finds that, treat the result as a Twitter API
error and don't cache the data. This should only happen if there is
some kind of error returned by Twitter.